refactor: change prefix `update` to `get`
