+++ src/server/constants/socket-actions.constants.ts
@@ -1,7 +1,7 @@
 
 export const enum GameActions {
-  UPDATE_USER_INFO = 'GAME:update-user-info',
-  UPDATE_BUSINESS_INFO = 'GAME:update-business-info',
+  GET_USER_INFO = 'GAME:get-user-info',
+  GET_BUSINESS_INFO = 'GAME:get-business-info',
   GET_BUSINESS_LIST = 'GAME:get-business-list',
   BUY_BUSINESS = 'GAME:buy-business',
   UPGRADE_BUSINESS = 'GAME:upgrade-business',
